+++ NEWS.md
@@ -4,6 +4,7 @@
 * 30 palettes have been added from ButterflyColors package. (#136)
 * 35 palettes have been added from MoMAColors package. (#142)
 * 32 palettes have been added from PrettyCols package. (#126)
+* 15 palettes have been added from waRhol package. (#131)
 
 # paletteer 1.5.0